pyproject.toml
[tool.poetry]
name = "to-do-list-viewer"
version = "0.1.0"
description = "Create tasks to manage your important life"
authors = ["Araslanov Roman <araslanovroman1991@gmail.com>"]
readme = "README.md"

packages = [
  { include = "data_forge" }
]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2.1"
loguru = "^0.7.3"
aiohttp = "^3.11.18"
uvicorn = "^0.34.2"
python-dotenv = "^1.1.0"
pydantic = "^2.11.4"
drf-spectacular = "^0.28.0"
drf-spectacular-sidecar = "^2025.5.1"
django-oauth-toolkit = "^3.0.1"
django-asgi-lifespan = "^0.4.0"
psycopg2-binary = "^2.9.10"
typer = "^0.15.4"
djangorestframework = "^3.16.0"
ruff = "^0.11.10"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
data_forge = "data_forge.cli:main"

[tool.black]
line-length = 88
target-version = ["py311"]


[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "C90", "PL", "B", "SIM", "RET", "ARG"]
fixable = ["ALL"]

